+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Added transformer for AWS DMS to CrateDB SQL
 
 ## 2024/07/19 v0.0.2
 - Added transformer for MongoDB CDC to CrateDB SQL conversion